+++ opencascade-hs/cpp/hs_types.h
@@ -43,6 +43,7 @@ typedef int BRepBuilderAPI_FaceError;
 typedef void BRepFilletAPI_MakeFillet;
 typedef void BRepFilletAPI_MakeChamfer;
 typedef void BRepPrimAPI_MakeBox;
+typedef void BRepPrimAPI_MakeTorus;
 typedef void TopExp_Explorer;
 typedef void BRepPrimAPI_MakeCylinder;
 typedef void BRepPrimAPI_MakeRevol;