+++ opencascade-hs/src/OpenCascade/BRepPrimAPI/Internal/Destructors.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE CApiFFI #-}
 module OpenCascade.BRepPrimAPI.Internal.Destructors 
 ( deleteMakeBox
+, deleteMakeTorus
 , deleteMakeRevol
 ) where
 
@@ -9,6 +10,7 @@ import OpenCascade.BRepPrimAPI.Types
 import Foreign.Ptr
 
 foreign import capi unsafe "hs_BRepPrimAPI_MakeBox.h hs_delete_BRepPrimAPI_MakeBox" deleteMakeBox :: Ptr MakeBox -> IO ()
+foreign import capi unsafe "hs_BRepPrimAPI_MakeTorus.h hs_delete_BRepPrimAPI_MakeTorus" deleteMakeTorus :: Ptr MakeTorus -> IO ()
 foreign import capi unsafe "hs_BRepPrimAPI_MakeRevol.h hs_delete_BRepPrimAPI_MakeRevol" deleteMakeRevol :: Ptr MakeRevol -> IO ()