+++ opencascade-hs/src/OpenCascade/BRepPrimAPI/Types.hs
@@ -3,6 +3,7 @@
 module OpenCascade.BRepPrimAPI.Types
 ( MakeBox
 , MakeRevol
+, MakeTorus
 ) where
 
 import OpenCascade.BRepBuilderAPI.MakeShape
@@ -10,6 +11,8 @@ import qualified OpenCascade.Inheritance as Inheritance
 
 data MakeBox
 data MakeRevol
+data MakeTorus
 
 instance Inheritance.SubTypeOf MakeShape MakeBox
 instance Inheritance.SubTypeOf MakeShape MakeRevol
+instance Inheritance.SubTypeOf MakeShape MakeTorus