+++ waterfall-cad/CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to the
 
 ## Unreleased
 
+### Added 
+
+- Added `Waterfall.Solids.torus`
+- Added Platonic Solids:
+    - `Waterfall.Solids.tetrahedron`
+    - `Waterfall.Solids.octahedron`
+    - `Waterfall.Solids.icosahedron`
+    - `Waterfall.Solids.dodecahedron`
+
 ## 0.6.0.1
 
 ### Added